Fix bug with angular-datepicker popping up on next year
For some reason there was this bit of code in the angular-datepicker source that made it so the year in the pop up was whatever year dateMaxLimit was in. So in this case I was trying to set dateMaxLimit to 6 months in the future and the calendar would pop up showing October 2016 rather than October 2015. Removing this bit of code fixes the problem and it doesn't seem to break anything else.